% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShapeRotator.R
\docType{data}
\name{double_data_1}
\alias{double_data_1}
\title{Landmark data from head-first burrowing worm lizards' crania}
\format{An array with landmark data (crania)}
\usage{
data(double_data_1)
}
\description{
Landmark data from head-first burrowing worm lizards' crania (Kazi & Hipsley, 2018)
}
\references{
S. Kazi & C.A. Hipsley. 2018. Conserved evolution of skull shape in Caribbean
head-first burrowing worm lizards (Squamata: Amphisbaenia) Biological Journal
of the Linnean Society 125:14-29

M. Vidal-García, L. Bandara and J.S. Keogh. 2018. ShapeRotator: An R tool
for standardized rigid rotations of articulated three-dimensional
structures with application for geometric morphometrics.
Ecology and Evolution. DOI: 10.1002/ece3.4018
}
\keyword{datasets}
