% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShapeRotator.R
\docType{data}
\name{simple_data_2}
\alias{simple_data_2}
\title{Landmark data from frog radioulnas}
\format{An array with landmark data (radioulnas)}
\usage{
data(simple_data_2)
}
\description{
Landmark data from frog radioulnas (Vidal-Garciía & Keogh, 2017)
}
\references{
M. Vidal-García & J.S. Keogh. 2017. Phylogenetic conservatism in skulls and
evolutionary lability in limbs–morphological evolution across an ancient frog
radiation is shaped by diet, locomotion and burrowing. BMC evolutionary biology 17:165

M. Vidal-García, L. Bandara and J.S. Keogh. 2018. ShapeRotator: An R tool
for standardized rigid rotations of articulated three-dimensional
structures with application for geometric morphometrics.
Ecology and Evolution. DOI: 10.1002/ece3.4018
}
\keyword{datasets}
